requirements.txt
python-telegram-bot>=21.0
openai>=1.0
pinecone-client>=3.0
python-dotenv>=1.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
